refactor(test) : 메서드 위치 변경에 따른 테스트 위치 변경
